:art: 로그아웃 및 회원탈퇴 UX수정
